Move extractPath to a separate module
