+++ modules/parsePath.js
@@ -1,19 +1,4 @@
-import warning from 'warning'
-
-export function extractPath(string, fullUrl = false) {
-  const match = string.match(/^https?:\/\/[^\/]*/)
-
-  if (match == null)
-    return string
-
-  warning(
-    fullUrl,
-    'A path must be pathname + search + hash only, not a fully qualified URL like "%s"',
-    string
-  )
-
-  return string.substring(match[0].length)
-}
+import extractPath from './extractPath'
 
 function parsePath(path) {
   let pathname = extractPath(path)
